fix: move network migrators
